Fin Ack implemented to end connection - Client side
Instead of sending a message over the connection to end it, Fin/ack is used to gracefully terminated the connection between host and client. Much better way to end a connection than sending over a string message, which could incidentially be replicated by the client.
This is done by the socket.shutdown method.